+++ modules/editors/default.nix
@@ -1,9 +1,8 @@
-{ ... }:
-
 {
   imports = [
     ./neovim.nix
     ./helix.nix
     ./vscode-server.nix
+    ./distributions
   ];
 }