+++ modules/editors/distributions/nvim/default.nix
@@ -1,0 +1,10 @@
+{ ... }:
+
+{
+  imports = [
+    ./lazyvim.nix
+    ./astronvim.nix
+    ./nvchad.nix
+    ./lunarvim.nix
+  ];
+}